Update attributes.rng
Change the sequence of conventionDeclarationReference, maintenanceEventReference, and sourceReference to be alphabetical. 

In this context, I've commented out the "attribute-group.lang-and-script-and-conventionDeclarationReference.optional" as I don't see where we would be using only conventionDeclarationReference, but not the other two assertion attributes, and moved the definition of conventionDeclarationReference into the section where the attribute-group.assertion-reference.optional is defined (i.e. same as for maintenanceEventReference and sourceReference.