Update models-and-shared-elements.rng
Re-added the "head" element, which was missing, though still referenced (at the moment) for EAC-CPF (generalContext structureOrGenealogy) as "element.head.optional"; while removed from those elements that include "head" next to the block formatting elements in EAD3, the element is still kept with "identificationData" and the numbered and unnumbered "c" elements.

On the other hand "num" is to be removed (same as the other mixed content element from EAD3) and replaced by "referringString". I have therefore removed the model "model.mixed-content-plus-num.optional".

Also included a comment that "abbreviation" can be deleted.